fix multi-region tests
